feat: create new support category and move msfs command to support category
